integrated basic user authentification

+++ src/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, UpdateResult, Repository } from 'typeorm';
 import { CreatePostDto } from './dto/create-post.dto';
@@ -13,37 +13,50 @@ export class PostService {
   ) {}
 
   async findAll(): Promise<Post[]> {
-    try {
-      return this.postsRepository.find();
-    } catch (err) {
-      console.log('Console: ' + err);
-    }
+    return await this.postsRepository.find();
   }
 
-  async findOneById(id: number): Promise<Post> {
-    return this.postsRepository.findOne({ where: { _id: id } });
+  async findOneById(id: string): Promise<Post> {
+    const post = await this.postsRepository.findOneByOrFail({ id });
+
+    return post;
   }
 
   async createOne(createPostDto: CreatePostDto): Promise<Post> {
+    if (!createPostDto.title || !createPostDto.text) {
+      throw new HttpException(
+        "Inputs shouldn't be empty!",
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     try {
       const post = new Post();
 
       post.title = createPostDto.title;
       post.text = createPostDto.text;
-      post.author = 111;
+      post.author = 'testAuthorId';
       post.createdAt = new Date();
       post.updatedAt = new Date();
 
       return this.postsRepository.save(post);
     } catch (err) {
+      console.log(err);
       return err;
     }
   }
 
   async updateOne(
-    id: number,
+    id: string,
     updatePostDto: UpdatePostDto,
   ): Promise<UpdateResult> {
+    if (!updatePostDto.title || !updatePostDto.text) {
+      throw new HttpException(
+        "Inputs shouldn't be empty!",
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     const updatedData = {
       title: updatePostDto.title,
       text: updatePostDto.text,
@@ -53,7 +66,13 @@ export class PostService {
     return this.postsRepository.update(id, updatedData);
   }
 
-  async deleteOne(id: number): Promise<DeleteResult> {
+  async deleteOne(id: string): Promise<DeleteResult> {
+    const post = await this.postsRepository.findOneByOrFail({ id });
+
+    if (!post) {
+      throw new HttpException('Post has not been found', HttpStatus.NOT_FOUND);
+    }
+
     return this.postsRepository.delete(id);
   }
 }
